Move sorting to Ordinare::Sort
